backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.1
sqlmodel==0.0.21
sqlalchemy==2.0.30
python-multipart==0.0.9
pillow==10.4.0
opencv-python==4.10.0.82
numpy==1.26.4
qrcode==7.4.2
pyjwt==2.9.0
colorthief==0.2.1
pandas==2.2.2
openpyxl==3.1.5
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
aiofiles==24.1.0
watchfiles==0.21.0
httpx==0.27.0

